requirements.txt
pyjwt==2.9.0
redis==5.0.8
motor==3.6.0
aiohttp==3.9.5 
async-lru==2.0.5
aioboto3==13.1.0
requests==2.32.3 
APScheduler==3.11.0
cryptography==42.0.8
python-telegram-bot==21.4

